+++ x/community/keeper/proposal_handler.go
@@ -19,22 +19,8 @@ func HandleCommunityPoolLendDepositProposal(ctx sdk.Context, k Keeper, p *types.
 
 // HandleCommunityPoolLendWithdrawProposal is a handler for executing a passed community pool lend withdraw proposal.
 func HandleCommunityPoolLendWithdrawProposal(ctx sdk.Context, k Keeper, p *types.CommunityPoolLendWithdrawProposal) error {
-	// hard allows attempting to withdraw more funds than there is a deposit for.
-	// this means the amount that gets withdrawn will not necessarily match the amount set in the proposal.
-	// to calculate how much is withdrawn, compare this module's balance before & after withdraw.
-	balanceBefore := k.bankKeeper.GetAllBalances(ctx, k.moduleAddress)
-
 	// withdraw funds from x/hard to this module account
-	err := k.hardKeeper.Withdraw(ctx, k.moduleAddress, p.Amount)
-	if err != nil {
-		return err
-	}
-
-	balanceAfter := k.bankKeeper.GetAllBalances(ctx, k.moduleAddress)
-	totalWithdrawn := balanceAfter.Sub(balanceBefore...)
-
-	// send all withdrawn coins back to community pool
-	return k.distrKeeper.FundCommunityPool(ctx, totalWithdrawn, k.moduleAddress)
+	return k.hardKeeper.Withdraw(ctx, k.moduleAddress, p.Amount)
 }
 
 // HandleCommunityCDPRepayDebtProposal is a handler for executing a passed community pool cdp repay debt proposal.